requirements.txt
python_telegram_bot==13.4.1
ffmpeg==1.4